SSOR-100 refactor for code reusability
