--- src/app/features/cards/slices/cardDetailSlice.js
@@ -1,81 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-import { getCardDetail, saveCardDetail } from "../../services/cardService";
-
-const initialState = {
-  card: {
-    id: "-",
-    numberHash: "-",
-    accountId: "-",
-    created: new Date(),
-    activeSince: new Date(),
-    expirationDate: new Date(),
-    pin: "-",
-    cvv: "-",
-    confirmed: "false",
-    active: "false",
-    stolen: "false",
-    cardType: "CARD_PLAIN",
-  },
-  props: {
-    isSavable: true,
-    isValidPin: true,
-    isValidCvv: true,
-    isValidNumberHash: true,
-    isValidAccountId: true,
-    isEditable: false,
-  },
-  status: "init",
-  error: null,
-};
-
-export const cardDetailSlice = createSlice({
-  name: "cardDetails",
-  initialState,
-  reducers: {
-    editCard(state, action) {
-      state.props.isEditable = action.payload;
-    },
-    validForm(state, action) {
-      state.props = action.payload;
-    },
-    updateCard(state, action) {
-      state.card = action.payload;
-    },
-  },
-  extraReducers: {
-    [getCardDetail.fulfilled]: (state, action) => {
-      let payload = action.payload;
-      payload.created = new Date(payload.created).toISOString();
-      payload.activeSince = new Date(payload.activeSince).toISOString();
-      payload.expirationDate = new Date(payload.expirationDate).toISOString();
-
-      state.card = payload;
-      state.status = action.meta.requestStatus;
-      state.error = null
-    },
-    [getCardDetail.pending]: (state, action) => {
-      state.status = action.meta.requestStatus;
-    },
-    [getCardDetail.rejected]: (state, action) => {
-      state.error = action.error.message;
-      state.status = action.error.name;
-    },
-    [saveCardDetail.fulfilled]: (state, action) => {
-      state.props.isEditable = false;
-      state.status = action.meta.requestStatus;
-      state.error = null
-    },
-    [saveCardDetail.pending]: (state, action) => {
-      state.status = action.meta.requestStatus;
-    },
-    [saveCardDetail.rejected]: (state, action) => {
-      state.error = action.error.message;
-      state.status = action.error.name;
-    },
-  },
-});
-
-export const { editCard, validForm, updateCard, canSave } =
-  cardDetailSlice.actions;
-
-export default cardDetailSlice.reducer;
