+++ src/app/store.js
@@ -1,8 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit";
 import loanSlice from "./features/loans/slice/loanSlice";
 
-import cardDetailReducer from "./features/cards/slices/cardDetailSlice";
-import cardCreateReducer from "./features/cards/slices/cardCreateSlice";
+import cardCreateReducer from "./features/cards/slices/card-create-slice";
+import cardDetailReducer from "./features/cards/slices/card-detail-slice";
 
 export const store = configureStore({
   reducer: {
